Components: Refactor into individual files
