+++ nightmare.vcxproj
@@ -120,7 +120,13 @@
     <ClCompile Include="src\atlas.cc" />
     <ClCompile Include="src\blob.cc" />
     <ClCompile Include="src\char.cc" />
-    <ClCompile Include="src\component\component.cc" />
+    <ClCompile Include="src\component\gas_production_component.cc" />
+    <ClCompile Include="src\component\light_component.cc" />
+    <ClCompile Include="src\component\power_component.cc" />
+    <ClCompile Include="src\component\relative_position.cc" />
+    <ClCompile Include="src\component\renderable_component.cc" />
+    <ClCompile Include="src\component\switchable_component.cc" />
+    <ClCompile Include="src\component\switch_component.cc" />
     <ClCompile Include="src\config.cc" />
     <ClCompile Include="src\input.cc" />
     <ClCompile Include="src\mesh.cc" />
@@ -146,7 +152,14 @@
     <ClInclude Include="src\char.h" />
     <ClInclude Include="src\chunk.h" />
     <ClInclude Include="src\common.h" />
-    <ClInclude Include="src\component\component.h" />
+    <ClInclude Include="src\component\component_manager.h" />
+    <ClInclude Include="src\component\gas_production_component.h" />
+    <ClInclude Include="src\component\light_component.h" />
+    <ClInclude Include="src\component\power_component.h" />
+    <ClInclude Include="src\component\relative_position.h" />
+    <ClInclude Include="src\component\renderable_component.h" />
+    <ClInclude Include="src\component\switchable_component.h" />
+    <ClInclude Include="src\component\switch_component.h" />
     <ClInclude Include="src\config.h" />
     <ClInclude Include="src\fixed_cube.h" />
     <ClInclude Include="src\fixed_grid.h" />
